fix(ci): python coverage shard - split oversized executor tests
